Safari-609.1.11.1/LayoutTests/fast/dom/document-all-tags.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test that HTMLAllCollection.tags() does not exist.');
shouldBeUndefined("document.all.tags");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
